fix(thermos): split main.py along helper boundaries (round-3 audit fixes)
main.py crossed the 1k-line ceiling at 1050 lines after the round-1/round-2
helper extractions. The wave plan's rubric-1 rule rejects that, and the
code-quality round-3 re-audit flagged it as a HIGH alongside the still-over-
300-line main() body. This commit moves the helpers to dedicated modules
along their existing boundaries.

* main.py: orchestrator only (~620 lines, well under the 1k-line ceiling)
* main_models.py: _resolve_run_budget + _resolve_agent_model +
  _first_runnable_in_chain + _resolve_model_chain_inputs
* main_setup.py: _run_setup_script + _compute_agent_deadline
* main_agent.py: _AgentRunArgs + _build_tool_context + _run_agent_with_timeout
  + the _run_agent_once / _execute_agent closures it owns
* main_outcome.py: _classify_outcome + _publish_span_attrs

Public surface preserved: __all__ retains MainResult, RunOutcome, main,
_ConfigurationError, _AgentTimeoutError. _ConfigurationError is still
importable from mergecraft.main (privilege.py consumes it).

Test-harness monkeypatch surface preserved: tests/support/run_main_harness.py
and the affected test files now patch resolve_model / resolve_runtime_agent
/ _first_runnable_in_chain / resolve_instructions on both mergecraft.main
AND the new helper modules in lock-step, since each helper module owns its
own import bindings.

Addresses round-3 code-quality findings:
- HIGH: main.py crossed the 1k-line ceiling (1050 -> ~620)
- HIGH: main() still 547 lines (target <= 300); now reads as a flat
  sequence of named helper calls with sub-helpers in dedicated modules
- MEDIUM: inline attrs_source lambda testing-incompatible
  -> _publish_span_attrs extracted to main_outcome.py
- LOW: _build_tool_context pass-through wrapper -> moved to main_agent.py
- LOW: unused lambda wrapper around _publish_span_attrs -> documented
  why it stays (tracer requires callable)
- LOW: hidden-by-reference mutations of tool_state -> comment added
- LOW: _AgentRunArgs.agent_id widens to str -> tightened to AgentId
- LOW: _resolve_agent_model docstring mis-describes boolean -> corrected
- LOW: _run_setup_script "stamps the failure reason on ToolState" claim
  -> corrected to "writes the reason; the orchestrator reads it"
- LOW: _AgentRunArgs "dozen" -> "fifteen"
- LOW: _classify_outcome "unknown policy" hedge -> closed vocabulary only
